pyproject.toml
[project]
name = "ptax-project-update"
version = "2.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "google-cloud-storage>=3.10.1",
    "polars>=1.21.0",
    "rq>=2.3.2",
    "usaddress>=0.5.11",
]

[dependency-groups]
dev = [
    "locust>=2.43.4",
]

[tool.pytest.ini_options]
timeout = 300
